requirements.txt
annotated-types==0.7.0
anyio==4.12.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
numpy==2.3.5
portalocker==3.2.0
protobuf==6.33.1
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.6
PyYAML==6.0.3
qdrant-client==1.16.1
requests==2.32.5
soupsieve==2.8
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
